Phase 6.8 — Extract test helpers into tests/helpers.ts (#113)
admin(), customerIds(), clearOrdersForWeek(), seedOrder() were
duplicated across six spec files with drifted shapes. Consolidated
into tests/helpers.ts:

- admin() — Supabase service-role client, exported alongside the
  existing adminClient alias.
- customerIds() — resolves the seeded test customers' UUIDs by
  token. Aliased as testCustomerIds (used by admin-send +
  notifications-flow).
- clearOrdersForWeek(weekOf) — deletes test-customer orders for a
  given week. Aliased as clearWeek (used by admin-orders +
  admin-orders-export).
- seedOrder(input) + SeedOrderInput — unions the supported fields
  (status?, needsReconciliation?) so all three caller specs share
  one implementation.

Updated 8 spec files to import from helpers and delete local copies.
Net −160 lines of duplicated helper code.
