+++ tests/admin-orders-export.spec.ts
@@ -1,7 +1,14 @@
 import { test, expect } from "@playwright/test";
-import { createClient } from "@supabase/supabase-js";
 
-import { ADMIN_STORAGE_STATE, TEST_CUSTOMERS, TEST_PRODUCTS } from "./helpers";
+import {
+  ADMIN_STORAGE_STATE,
+  TEST_CUSTOMERS,
+  TEST_PRODUCTS,
+  admin,
+  customerIds,
+  clearWeek,
+  seedOrder,
+} from "./helpers";
 import { weekOfMondayNY } from "@/lib/week";
 import {
   EXPORT_COLUMNS,
@@ -10,68 +17,6 @@ import {
   toTsv,
 } from "@/lib/admin/export-orders";
 import type { OrderRow } from "@/lib/admin/orders-queries";
-
-function admin() {
-  return createClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.SUPABASE_SERVICE_ROLE_KEY!,
-    { auth: { autoRefreshToken: false, persistSession: false } },
-  );
-}
-
-async function customerIds(): Promise<{ farmStand: string; restaurant: string }> {
-  const sb = admin();
-  const { data, error } = await sb
-    .from("customers")
-    .select("id, token")
-    .in("token", [TEST_CUSTOMERS.farmStand.token, TEST_CUSTOMERS.restaurant.token]);
-  if (error || !data) throw new Error(`customerIds: ${error?.message ?? "no rows"}`);
-  const map = new Map(data.map((r) => [r.token, r.id]));
-  return {
-    farmStand: map.get(TEST_CUSTOMERS.farmStand.token)!,
-    restaurant: map.get(TEST_CUSTOMERS.restaurant.token)!,
-  };
-}
-
-async function clearWeek(weekOf: string): Promise<void> {
-  const sb = admin();
-  const ids = await customerIds();
-  await sb
-    .from("orders")
-    .delete()
-    .in("customer_id", [ids.farmStand, ids.restaurant])
-    .eq("week_of", weekOf);
-}
-
-type SeedOrderInput = {
-  customerId: string;
-  weekOf: string;
-  fulfillmentType: "pickup" | "delivery";
-  items: Array<{ productId: string; qty: number; unitPriceCents: number }>;
-};
-
-async function seedOrder(input: SeedOrderInput): Promise<string> {
-  const sb = admin();
-  const { data: order } = await sb
-    .from("orders")
-    .insert({
-      customer_id: input.customerId,
-      week_of: input.weekOf,
-      fulfillment_type: input.fulfillmentType,
-      status: "new",
-    })
-    .select("id")
-    .single();
-  await sb.from("order_items").insert(
-    input.items.map((i) => ({
-      order_id: order!.id,
-      product_id: i.productId,
-      qty: i.qty,
-      unit_price_cents: i.unitPriceCents,
-    })),
-  );
-  return order!.id;
-}
 
 // --- unit-style tests (no `page`) ----------------------------------------
