+++ tests/helpers.ts
@@ -30,12 +30,91 @@ export const TEST_PRODUCTS = {
 } as const;
 
 // Lazy admin Supabase client — only constructed when a test calls it.
-function adminClient() {
+// Exported as both `admin` (the common shorthand used in admin-side specs)
+// and `adminClient` (the name customer-side specs picked up). Same function;
+// two names so the Phase 6.8 consolidation didn't churn every existing call.
+export function admin() {
   const url = process.env.NEXT_PUBLIC_SUPABASE_URL!;
   const key = process.env.SUPABASE_SERVICE_ROLE_KEY!;
   return createClient(url, key, {
     auth: { autoRefreshToken: false, persistSession: false },
   });
+}
+export { admin as adminClient };
+
+// Resolves the two seeded test customers' UUIDs by token. Used by every spec
+// that needs to query orders/sends/etc. scoped to the test fixtures.
+export async function customerIds(): Promise<{ farmStand: string; restaurant: string }> {
+  const sb = admin();
+  const { data, error } = await sb
+    .from("customers")
+    .select("id, token")
+    .in("token", [TEST_CUSTOMERS.farmStand.token, TEST_CUSTOMERS.restaurant.token]);
+  if (error || !data) throw new Error(`customerIds: ${error?.message ?? "no rows"}`);
+  const map = new Map(data.map((r) => [r.token, r.id]));
+  return {
+    farmStand: map.get(TEST_CUSTOMERS.farmStand.token)!,
+    restaurant: map.get(TEST_CUSTOMERS.restaurant.token)!,
+  };
+}
+// admin-send + notifications-flow specs settled on `testCustomerIds`; alias
+// rather than churn every call site.
+export { customerIds as testCustomerIds };
+
+// Deletes orders for the two seeded test customers in a given NY-week-Monday.
+// Caller passes the week explicitly (use `weekOfMondayNY()` from @/lib/week
+// for the current week). admin-orders + admin-orders-export specs called
+// this `clearWeek`; alias preserved for the same reason.
+export async function clearOrdersForWeek(weekOf: string): Promise<void> {
+  const sb = admin();
+  const ids = await customerIds();
+  await sb
+    .from("orders")
+    .delete()
+    .in("customer_id", [ids.farmStand, ids.restaurant])
+    .eq("week_of", weekOf);
+}
+export { clearOrdersForWeek as clearWeek };
+
+// Inserts an order + its line items for a seeded test customer. Union of the
+// shapes used across admin-orders, admin-orders-export, and orders-flow specs
+// (status, needsReconciliation are optional so any caller can omit them).
+export type SeedOrderInput = {
+  customerId: string;
+  weekOf: string;
+  fulfillmentType: "pickup" | "delivery";
+  status?: "new" | "ready" | "picked-up" | "delivered";
+  needsReconciliation?: boolean;
+  items: Array<{ productId: string; qty: number; unitPriceCents: number }>;
+};
+export async function seedOrder(input: SeedOrderInput): Promise<string> {
+  const sb = admin();
+  const { data: order, error: oErr } = await sb
+    .from("orders")
+    .insert({
+      customer_id: input.customerId,
+      week_of: input.weekOf,
+      fulfillment_type: input.fulfillmentType,
+      delivery_address:
+        input.fulfillmentType === "delivery" ? "123 Test St" : null,
+      delivery_preference:
+        input.fulfillmentType === "delivery" ? "Front door" : null,
+      status: input.status ?? "new",
+      needs_reconciliation: input.needsReconciliation ?? false,
+    })
+    .select("id")
+    .single();
+  if (oErr || !order) throw new Error(`seedOrder: ${oErr?.message}`);
+
+  const rows = input.items.map((i) => ({
+    order_id: order.id,
+    product_id: i.productId,
+    qty: i.qty,
+    unit_price_cents: i.unitPriceCents,
+  }));
+  const { error: iErr } = await sb.from("order_items").insert(rows);
+  if (iErr) throw new Error(`seedOrder items: ${iErr.message}`);
+  return order.id;
 }
 
 // Clears current-NY-week orders for the seeded test customers and restores
@@ -49,7 +128,7 @@ function adminClient() {
 // (a delivered 2026-04-27 prior order used to assert delivery_preference
 // prefill) is preserved.
 export async function resetCustomerOrderState(): Promise<void> {
-  const sb = adminClient();
+  const sb = admin();
   const currentWeek = weekOfMondayNY();
   for (const c of Object.values(TEST_CUSTOMERS)) {
     const { data } = await sb
@@ -83,7 +162,7 @@ export async function resetCustomerOrderState(): Promise<void> {
 // state tests. The schedule table has a single row by design; updating any
 // row updates the schedule.
 export async function setOrderingOpen(open: boolean): Promise<void> {
-  const sb = adminClient();
+  const sb = admin();
   await sb
     .from("ordering_schedule")
     .update({ is_open: open })
@@ -102,7 +181,7 @@ export async function setOrderingOpen(open: boolean): Promise<void> {
 // devs run tests against the same dev DB in parallel.
 export type ProductQtySnapshot = Array<{ id: string; qty_available: number }>;
 export async function setAllProductsSoldOut(): Promise<ProductQtySnapshot> {
-  const sb = adminClient();
+  const sb = admin();
   const { data } = await sb
     .from("products")
     .select("id, qty_available")
@@ -117,7 +196,7 @@ export async function setAllProductsSoldOut(): Promise<ProductQtySnapshot> {
 }
 
 export async function restoreProductQty(snapshot: ProductQtySnapshot): Promise<void> {
-  const sb = adminClient();
+  const sb = admin();
   for (const row of snapshot) {
     await sb
       .from("products")
@@ -130,6 +209,6 @@ export async function restoreProductQty(snapshot: ProductQtySnapshot): Promise<v
 // per-item sold-out) call this; cleanup is via resetCustomerOrderState in
 // afterEach, which only knows about TEST_PRODUCTS — pass an id in that set.
 export async function setProductQty(id: string, qty: number): Promise<void> {
-  const sb = adminClient();
+  const sb = admin();
   await sb.from("products").update({ qty_available: qty }).eq("id", id);
 }
